Pin lodash-node below 3.9
`lodash-node 3.9` breaks in the browser (it wants `root` to exist).  A temporary fix is to pin the version below 3.9 (3.8.x).  The real fix is to change the deps to `lodash` instead of `lodash-node`.
 
`lodash-node` is deprecated and, was never intended to be used in the browser in the first place.  It's been working by accident up to this version.